bus_booking_backend/requirements.txt
Django==6.0.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
python-dotenv>=1.0.0
Pillow>=10.1.0
gunicorn>=21.2.0
requests>=2.31.0
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
